feat: add centralized chat state management hook and related types
- Implemented `useChatState` hook for managing AI chat state, including refs, state, and derived values.
- Created `chatTypes.ts` to define event payload types for AI chat interactions.
- Added an entry point for the agent module in `index.ts` to facilitate exports.
